Cover all categories in place page golden test
