+++ internal/server/v1/page/golden/place_page/dummy.json.Overview.json
@@ -1,4 +1,7 @@
 {
+  "statVarSeries": {
+    "dummy": {}
+  },
   "validCategories": {
     "dummy": {}
   }
