+++ internal/server/v1/page/place_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package placepage
+package page
 
 import (
 	"testing"
